+++ spec/unit/validator_spec.rb
@@ -31,7 +31,7 @@ module Pod
     #
     def stub_podspec(pattern = nil, replacement = nil)
       spec = (fixture('spec-repos') + 'master/Specs/JSONKit/1.4/JSONKit.podspec.json').read
-      spec.gsub!(/https:\/\/github\.com\/johnezang\/JSONKit\.git/, fixture('integration/JSONKit').to_s)
+      spec.gsub!(%r{https://github\.com/johnezang/JSONKit\.git}, fixture('integration/JSONKit').to_s)
       spec.gsub!(pattern, replacement) if pattern && replacement
       spec
     end
@@ -322,7 +322,9 @@ module Pod
         it 'includes the use_frameworks!(false) directive' do
           podfile = @validator.send(:podfile_from_spec, :ios, '5.0', false)
           target_definition = podfile.target_definitions['Pods']
+          # rubocop:disable Style/DoubleNegation
           (!!target_definition.uses_frameworks?).should == false
+          # rubocop:enable Style/DoubleNegation
         end
       end
 
@@ -334,10 +336,21 @@ module Pod
         podfile.sources.should == sources
       end
 
-      it 'uses xcodebuild to generate notes and warnings' do
+      it 'uses xcodebuild to generate warnings' do
         validator = Validator.new(podspec_path, SourcesManager.master.map(&:url))
         validator.stubs(:check_file_patterns)
-        validator.stubs(:xcodebuild).returns("file.m:1:1: warning: direct access to objective-c's isa is deprecated")
+        validator.stubs(:xcodebuild).returns("file.m:1:1: warning: 'dataFromPropertyList:format:errorDescription:' is deprecated: first deprecated in iOS 8.0 - Use dataWithPropertyList:format:options:error: instead. [-Wdeprecated-declarations]")
+        validator.stubs(:validate_url)
+        validator.validate
+        first = validator.results.map(&:to_s).first
+        first.should.include '[xcodebuild]'
+        validator.result_type.should == :warning
+      end
+
+      it 'uses xcodebuild to generate notes' do
+        validator = Validator.new(podspec_path, SourcesManager.master.map(&:url))
+        validator.stubs(:check_file_patterns)
+        validator.stubs(:xcodebuild).returns("file.m:1:1: note: 'dataFromPropertyList:format:errorDescription:' has been explicitly marked deprecated here")
         validator.stubs(:validate_url)
         validator.validate
         first = validator.results.map(&:to_s).first